size_t fix

+++ matrix/src/matrix.cpp
@@ -1,5 +1,7 @@
+//
+// Created by ACER on 10.10.2020.
+//
 #include "matrix.h"
-
 Matrix::Matrix() : row(1), col(1) {
     allocateMemory();
     markWithOnes();
@@ -12,15 +14,15 @@ Matrix::Matrix(size_t rows, size_t cols) : row(rows), col(cols) {
 
 Matrix::Matrix(const Matrix &copy) : row(copy.row), col(copy.col) {
     allocateMemory();
-    for (int i = 0; i < row; ++i) {
-        for (int j = 0; j < col; ++j) {
+    for (size_t i = 0; i < row; ++i) {
+        for (size_t j = 0; j < col; ++j) {
             arr[i][j] = copy[i][j];
         }
     }
 }
 
 Matrix::~Matrix() {
-    for (int i = 0; i < row; ++i) {
+    for (size_t i = 0; i < row; ++i) {
         delete[] arr[i];
     }
     delete[] arr;
@@ -40,8 +42,8 @@ Matrix &Matrix::operator=(const Matrix &a) {
     col = a.getCol();
     allocateMemory();
 
-    for (int i = 0; i < row; ++i) {
-        for (int j = 0; j < col; ++j) {
+    for (size_t i = 0; i < row; ++i) {
+        for (size_t j = 0; j < col; ++j) {
             arr[i][j] = a[i][j];
         }
     }
@@ -79,13 +81,13 @@ void Matrix::set(size_t n, size_t m, const double &value) {
 
 void Matrix::resize(size_t n, size_t m) {
     auto *temp = new Matrix(n, m);
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < m; ++j) {
+    for (size_t i = 0; i < n; ++i) {
+        for (size_t j = 0; j < m; ++j) {
             temp->arr[i][j] = 0;
         }
     }
-    for (int i = 0; i < getMin(n, row); ++i) {
-        for (int j = 0; j < getMin(m, col); ++j) {
+    for (size_t i = 0; i < getMin(n, row); ++i) {
+        for (size_t j = 0; j < getMin(m, col); ++j) {
             temp->arr[i][j] = arr[i][j];
         }
     }
@@ -97,24 +99,24 @@ void Matrix::resize(size_t n, size_t m) {
 
 void Matrix::allocateMemory() {
     arr = new double *[row];
-    for (int i = 0; i < row; ++i) {
+    for (size_t i = 0; i < row; ++i) {
         arr[i] = new double[col];
     }
 }
 
 void Matrix::markWithOnes() {
-    for (int i = 0; i < row; ++i) {
-        for (int j = 0; j < col; ++j) {
+    for (size_t i = 0; i < row; ++i) {
+        for (size_t j = 0; j < col; ++j) {
             arr[i][j] = (i == j ? 1 : 0);
         }
     }
 }
 
-int Matrix::getMin(int a, int b) {
+size_t Matrix::getMin(size_t a, size_t b) {
     return (a > b ? b : a);
 }
 
-void Matrix::checkMismatch(int n, int m) const {
+void Matrix::checkMismatch(size_t n, size_t m) const {
     if (n != row || m != col) {
         throw SizeMismatchException();
     }
@@ -128,8 +130,8 @@ void Matrix::checkMismatch() const {
 
 Matrix &Matrix::operator+=(const Matrix &a) {
     checkMismatch(a.row, a.col);
-    for (int i = 0; i < row; ++i) {
-        for (int j = 0; j < col; ++j) {
+    for (size_t i = 0; i < row; ++i) {
+        for (size_t j = 0; j < col; ++j) {
             arr[i][j] += a[i][j];
         }
     }
@@ -152,8 +154,8 @@ Matrix Matrix::operator-(const Matrix &a) const {
 
 Matrix operator*(const double &a, const Matrix &b) {
     Matrix temp = b;
-    for (int i = 0; i < temp.getRow(); ++i) {
-        for (int j = 0; j < temp.getCol(); ++j) {
+    for (size_t i = 0; i < temp.getRow(); ++i) {
+        for (size_t j = 0; j < temp.getCol(); ++j) {
             temp[i][j] *= a;
         }
     }
@@ -171,8 +173,8 @@ Matrix Matrix::operator+() const {
 }
 
 Matrix &Matrix::operator*=(const double &number) {
-    for (int i = 0; i < row; ++i) {
-        for (int j = 0; j < col; ++j) {
+    for (size_t i = 0; i < row; ++i) {
+        for (size_t j = 0; j < col; ++j) {
             arr[i][j] *= number;
         }
     }
@@ -189,10 +191,10 @@ Matrix &Matrix::operator*=(const Matrix &a) {
         throw SizeMismatchException();
     }
     Matrix temp(row, a.getCol());
-    for (int i = 0; i < row; ++i) {
-        for (int j = 0; j < temp.getCol(); ++j) {
+    for (size_t i = 0; i < row; ++i) {
+        for (size_t j = 0; j < temp.getCol(); ++j) {
             double s = 0;
-            for (int k = 0; k < col; ++k) {
+            for (size_t k = 0; k < col; ++k) {
                 s += arr[i][k] * a[k][j];
             }
             temp[i][j] = s;
@@ -208,8 +210,8 @@ Matrix Matrix::operator*(const Matrix &a) const {
 
 bool Matrix::operator==(const Matrix &a) const {
     checkMismatch();
-    for (int i = 0; i < row; ++i) {
-        for (int j = 0; j < col; ++j) {
+    for (size_t i = 0; i < row; ++i) {
+        for (size_t j = 0; j < col; ++j) {
             if (!areClose(arr[i][j], a[i][j])) {
                 return false;
             }
@@ -232,7 +234,7 @@ double Matrix::abs(double a) {
 
 std::vector<double> Matrix::getRow(size_t row) {
     std::vector<double> vec(getRow());
-    for (int i = 0; i < getRow(); ++i) {
+    for (size_t i = 0; i < getRow(); ++i) {
         vec[i] = arr[row][i];
     }
     return vec;
@@ -240,15 +242,15 @@ std::vector<double> Matrix::getRow(size_t row) {
 
 std::vector<double> Matrix::getColumn(size_t column) {
     std::vector<double> vec(getCol());
-    for (int i = 0; i < getCol(); ++i) {
+    for (size_t i = 0; i < getCol(); ++i) {
         vec[i] = arr[i][column];
     }
     return vec;
 }
 
 std::ostream &operator<<(std::ostream &output, const Matrix &matrix) {
-    for (int i = 0; i < matrix.getRow(); ++i) {
-        for (int j = 0; j < matrix.getCol(); ++j) {
+    for (size_t i = 0; i < matrix.getRow(); ++i) {
+        for (size_t j = 0; j < matrix.getCol(); ++j) {
             output << matrix[i][j];
         }
     }
@@ -256,11 +258,11 @@ std::ostream &operator<<(std::ostream &output, const Matrix &matrix) {
 }
 
 std::istream &operator>>(std::istream &input, Matrix &matrix) {
-    int n, m;
+    size_t n, m;
     input >> n >> m;
     matrix.resize(n, m);
-    for (int i = 0; i < n; ++i) {
-        for (int j = 0; j < m; ++j) {
+    for (size_t i = 0; i < n; ++i) {
+        for (size_t j = 0; j < m; ++j) {
             input >> matrix[i][j];
         }
     }
@@ -270,16 +272,16 @@ std::istream &operator>>(std::istream &input, Matrix &matrix) {
 double Matrix::det() const {
     checkMismatch();
     Matrix arr = *this;
-    for (int i = 0; i < arr.getRow() - 1; ++i) {
-        for (int j = i + 1; j < arr.getCol(); ++j) {
+    for (size_t i = 0; i < arr.getRow() - 1; ++i) {
+        for (size_t j = i + 1; j < arr.getCol(); ++j) {
             double coef = double(arr[j][i]) / arr[i][i];
-            for (int k = i; k < arr.getRow(); ++k) {
+            for (size_t k = i; k < arr.getRow(); ++k) {
                 arr[j][k] -= arr[i][k] * coef;
             }
         }
     }
     double d = 1;
-    for (int i = 0; i < arr.getRow(); ++i) {
+    for (size_t i = 0; i < arr.getRow(); ++i) {
         d *= arr[i][i];
     }
     return d;
@@ -288,7 +290,7 @@ double Matrix::det() const {
 double Matrix::trace() const {
     checkMismatch();
     double s = 0;
-    for (int i = 0; i < getRow(); ++i) {
+    for (size_t i = 0; i < getRow(); ++i) {
         s += arr[i][i];
     }
     return s;
@@ -296,8 +298,8 @@ double Matrix::trace() const {
 
 Matrix Matrix::transposed() const {
     Matrix temp(getCol(), getRow());
-    for (int i = 0; i < getRow(); ++i) {
-        for (int j = 0; j < getCol(); ++j) {
+    for (size_t i = 0; i < getRow(); ++i) {
+        for (size_t j = 0; j < getCol(); ++j) {
             temp[j][i] = arr[i][j];
         }
     }
